Add support to auto-accept any host fingerprint
In some embedded environments the hostkeys may change while the server IP/name keeps the same. In order not to need to manually connect to each device before being able to use pscp/plink in a batch, a mechanism is added to automatically extract the host fingerprint from the error message and try to reconnect using that key for the current session. This feature is enabled by the setting `"hostkey": "*"`. This allows setting specific keys as well.